Move AdditionalInfo interface to it's own file
